+++ include/engine/api/route_api.hpp
@@ -68,6 +68,7 @@ class RouteAPI : public BaseAPI
         response.values["waypoints"] = BaseAPI::MakeWaypoints(all_start_end_points);
         response.values["routes"] = std::move(jsRoutes);
         response.values["code"] = "Ok";
+        response.values["data_version"] = facade.GetTimestamp();
     }
 
   protected: